refactor: removed LoginForm component
